requirements.txt
PySide6>=6.5.0
python-dotenv>=1.0.0
transformers>=4.40.0
torch>=2.2.0